create shared connection flow component
