+++ packages/desktop/views/dashboard/drawers/dapp-config/views/index.ts
@@ -6,3 +6,6 @@ export { default as InputConnectionCodeDrawer } from './InputConnectionCodeDrawe
 export { default as EditAccountsDrawer } from './EditAccountsDrawer.svelte'
 export { default as EditNetworksDrawer } from './EditNetworksDrawer.svelte'
 export { default as EditPermissionsDrawer } from './EditPermissionsDrawer.svelte'
+export { default as OneClickAuthFlow } from './OneClickAuthFlow.svelte'
+export { default as SessionProposalFlow } from './SessionProposalFlow.svelte'
+export { default as SessionProposalFlowShared } from './SessionProposalFlowShared.svelte'
